requirements.txt
librosa>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
jupyter>=1.0.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
datasets>=2.0.0
flask>=2.0.0
xgboost>=1.6.0
lightgbm>=3.3.0
gradio>=4.0.0
seaborn>=0.11.0
pyyaml>=6.0
soundfile>=0.12.0
torchaudio>=0.12.0
wandb>=0.13.0
mlflow>=2.0.0
